requirements.txt
streamlit
praw
vaderSentiment
mysql-connector-python
matplotlib
wordcloud
pandas